Handons4/jobs/job1/requirements.txt
pytesseract
pillow
pandas
openpyxl
